refactor(persistAtomEffect): 재사용 가능한 effect 분리 및 디렉터리 변경
